\name{colorpalette}
\alias{colorpalette}
\title{Colorpalette}
\description{provides color palettes of several R packages (grDevices, RColorBrewer, colorRamps and own palettes)}
\usage{
colorpalette(pal, nrcol = NULL, alpha = NULL)
}
\arguments{
  \item{pal}{\code{which colorpalette should be chosen ( see disco() )}}
  \item{nrcol}{\code{number of colors to be used}}
  \item{alpha}{\code{alpha value for color opacity}}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{disco}},\code{\link{display.colorpalette}},\code{\link{LSD.demo.tour}}}
\examples{disco()}
\keyword{color}
