\name{convertcolor}
\alias{convertcolor}
\alias{demo.convertcolor}
\title{Convertcolor}
\description{convert R colors to hexadecimal representation}
\usage{
convertcolor(color, alpha = NULL)
}
\arguments{
  \item{color}{\code{vector of colors}}
  \item{alpha}{\code{alpha value for color opacity}}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{colorpalette}},\code{\link{convertgrey}},\code{\link{LSD.demo.tour}}}
\examples{demo.convertcolor()}
\keyword{color}
