\name{convertgrey}
\alias{convertgrey}
\alias{demo.convertgrey}
\title{Convertgrey}
\description{convert colorpalette to grayscales}
\usage{
convertgrey(color, alpha = NULL)
}
\arguments{
  \item{color}{\code{vector of colors}}
  \item{alpha}{\code{alpha value for color opacity}}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{colorpalette}},\code{\link{convertcolor}},\code{\link{LSD.demo.tour}}}
\examples{demo.convertgrey}
\keyword{color}
