\name{disco}
\alias{disco}
\alias{display.all.colorpalette}
\title{Disco}
\description{displays color palettes of several R packages (grDevices, RColorBrewer, colorRamps and own palettes)}
\usage{
disco(nr = 11)

display.all.colorpalette(nr = 11)
}
\arguments{
  \item{nr}{\code{number of colors to be used}}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{display.colorpalette}},\code{\link{distinctcolors}},\code{\link{LSD.demo.tour}}}
\examples{disco()}
\keyword{color}
