\name{LSD-package}
\alias{LSD-package}
\alias{LSD}
\docType{package}
\title{Lots of Superior Depictions}
\description{Create lots of colorful plots in a plethora of variations ( try the LSD.demo.tour() )}
\details{
\tabular{ll}{
Package: \tab LSD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-06\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bjoern Schwalb, Achim Tresch, Romain Francois

Maintainer: Bjoern Schwalb <schwalb@lmb.uni-muenchen.de>
}
\keyword{hplot}
\keyword{color}
\examples{LSD.demo.tour()}
