% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.utilities.R
\name{webdesign}
\alias{webdesign}
\alias{LSD.webdesign}
\title{Colored rectangular grid}
\usage{
webdesign(xlim, ylim, colpal = "rdbu", xlabels = NULL, ylabels = NULL, ...)
}
\arguments{
\item{xlim}{x limits, standard graphics parameter.}

\item{ylim}{y limits, standard graphics parameter.}

\item{colpal}{a character vector containing R built-in color names or a name of a \code{LSD} colorpalette as a character string (see disco() or \code{\link{disco}}).}

\item{xlabels}{a character vector containing labels depicted parallel to the x-axis.}

\item{ylabels}{a character vector containing labels depicted parallel to the y-axis.}

\item{...}{additional parameters to be passed to abline().}
}
\description{
Adds a colored rectangular grid to an existing plot.
}
\examples{
emptyplot(c(-5,5),c(-5,5))
labels = c("2 fold","4 fold","8 fold")
webdesign(c(-5,5),c(-5,5),lty = 2,xlabels = labels,ylabels = labels)
}
\seealso{
\code{\link{demotour}}, \code{\link{disco}}, \code{\link{colorpalette}}
}
\author{
Bjoern Schwalb
}
\keyword{grid}
\keyword{web}
