\name{windowxy}
\alias{windowxy}
\alias{demo.windowxy}
\title{Window xy}
\description{Creates a factorization of the number of windows for plots with device partitions to be used in par(mfrow = ...).}
\usage{
windowxy(windows = 1)
}
\arguments{
  \item{windows}{\code{number of windows}}
}
\author{Bjoern Schwalb, Achim Tresch}
\seealso{\code{\link{LSD.demo.tour}}}
\examples{demo.windowxy()}
\keyword{dplot}
