\name{colorpalette}
\alias{colorpalette}
\title{Colorpalette}
\description{provides color palettes of several R packages (grDevices, RColorBrewer, colorRamps and own palettes)}
\usage{
colorpalette(pal, nrcol = NULL, alpha = NULL)
}
\arguments{
  \item{pal}{which colorpalette should be chosen ( see disco() )}
  \item{nrcol}{number of colors to be used}
  \item{alpha}{alpha value for color opacity}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{disco}},\code{\link{display.colorpalette}},\code{\link{LSD.demo.tour}}}
\examples{disco()}
\keyword{color}
