% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LSD.demotour.R
\name{demotour}
\alias{LSD.demotour}
\alias{demotour}
\title{LSD teaser}
\usage{
demotour()
}
\description{
A compilation of selected plot examples.
}
\examples{
demotour()
}
\author{
Bjoern Schwalb
}
\seealso{
\code{\link{heatscatter}}, \code{\link{clusterplot}}, \code{\link{disco}}, \code{\link{colorpalette}}, \code{\link{daltonize}}
}
\keyword{clusterplot,}
\keyword{colorpalette,}
\keyword{daltonize}
\keyword{disco,}
\keyword{heatscatter,}

