\name{display.colorpalette}
\alias{display.colorpalette}
\alias{demo.display.colorpalette}
\title{Display colorpalette}
\description{displays color palettes of several R packages (grDevices, RColorBrewer, colorRamps and own palettes)}
\usage{
display.colorpalette(name, n = NULL, alpha = NULL)
}
\arguments{
  \item{name}{\code{name of the colorpalette}}
  \item{n}{\code{number of colors to be used}}
  \item{alpha}{\code{alpha value for color opacity}}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{disco}},\code{\link{distinctcolors}},\code{\link{LSD.demo.tour}}}
\examples{demo.display.colorpalette()}
\keyword{color}
