% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.utilities.R
\name{emptyplot}
\alias{emptyplot}
\alias{LSD.emptyplot}
\title{Wrapper function for an empty graphics device}
\usage{
emptyplot(xlim = c(-1, 1), ylim = c(-1, 1), ...)
}
\arguments{
\item{xlim}{x limits, standard graphics parameter.}

\item{ylim}{y limits, standard graphics parameter.}

\item{...}{additional parameters to be passed to points and plot.}
}
\description{
Calls an empty graphics device with a coordinate system of choice.
}
\examples{
emptyplot()
}
\seealso{
\code{\link{demotour}}
}
\author{
Bjoern Schwalb
}
\keyword{empty}
\keyword{plotting}
\keyword{region}
