\name{gridfkt}
\alias{gridfkt}
\alias{demo.gridfkt}
\title{Gridfunction}
\description{produces a colored grid for quadratic plots with log2 folds}
\usage{
gridfkt(lim, lty = 2, lwd = 1, addlines = TRUE, folds = TRUE)
}
\arguments{
  \item{lim}{limits for quadratic plots}
  \item{lty}{linetype to be used}
  \item{lwd}{linewidth to be used}
  \item{addlines}{should fold lines be added}
  \item{folds}{should labels be added to fold lines}
}
\author{Bjoern Schwalb}
\seealso{\code{\link{LSD.demo.tour}}}
\examples{demo.gridfkt()}
\keyword{aplot}
