% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LSD.utilities.R
\name{windowxy}
\alias{LSD.windowxy}
\alias{windowxy}
\title{Factorization of the number of windows for plots with device partitions}
\usage{
windowxy(nrwin = 1)
}
\arguments{
\item{nrwin}{a non-negative integer specifying the number of windows.}
}
\description{
Create a factorization of the number of windows for plots with device partitions to be used in par(mfrow = ...).
}
\examples{
windowxy(20)
}
\author{
Bjoern Schwalb
}
\seealso{
\code{\link{demotour}}
}
\keyword{factorization}

